+++ backend/integration_tests/src/communities/join_channel_tests.rs
@@ -75,6 +75,7 @@ fn join_private_channel_fails() {
             community_id,
             channel_id,
             invite_code: None,
+            referred_by: None,
             verified_credential_args: None,
         },
     );
@@ -153,6 +154,7 @@ fn join_community_and_channel_in_single_call_succeeds() {
             community_id,
             channel_id,
             invite_code: None,
+            referred_by: None,
             verified_credential_args: None,
         },
     );
@@ -207,6 +209,7 @@ fn invite_non_community_member_to_channel_succeeds() {
             community_id,
             channel_id,
             invite_code: None,
+            referred_by: None,
             verified_credential_args: None,
         },
     );